backend/requirement.txt
flask
gunicorn
numpy
opencv-python
flask-cors
pillow
face-recognition
